requirements.txt
Flask
google-generativeai
python-dotenv
google-cloud-texttospeech
